+++ R/form.R
@@ -149,11 +149,11 @@ format.select <- function(x, ...) {
 
 parse_options <- function(options) {
   parse_option <- function(option) {
-    attr <- as.list(xml2::xml_attrs(option))
+    name <- xml2::xml_text(option)
     list(
-      value = attr$value,
-      name = xml2::xml_text(option),
-      selected = !is.null(attr$selected)
+      value = xml2::xml_attr(option, "value", default = name),
+      name = name,
+      selected = xml2::xml_has_attr(option, "selected")
     )
   }